Added MongoDB BusinessCard mgr (et al)

+++ phoss-smp-backend-sql/src/main/java/com/helger/phoss/smp/backend/sql/mgr/SMPBusinessCardManagerSQL.java
@@ -358,7 +358,7 @@ public final class SMPBusinessCardManagerSQL extends AbstractSMPJPAEnabledManage
   }
 
   @Nonnegative
-  public int getSMPBusinessCardCount ()
+  public long getSMPBusinessCardCount ()
   {
     JPAExecutionResult <Number> ret;
     ret = doInTransaction ( () -> {
