+++ phoss-smp-backend/src/test/java/com/helger/phoss/smp/mock/MockSMPBusinessCardManager.java
@@ -48,7 +48,7 @@ final class MockSMPBusinessCardManager implements ISMPBusinessCardManager
     throw new UnsupportedOperationException ();
   }
 
-  public int getSMPBusinessCardCount ()
+  public long getSMPBusinessCardCount ()
   {
     return 0;
   }
