+++ phoss-smp-webapp/src/test/java/com/helger/phoss/smp/domain/businesscard/ISMPBusinessCardManagerFuncTest.java
@@ -96,7 +96,7 @@ public final class ISMPBusinessCardManagerFuncTest
       ISMPBusinessCard aBusinessCard = null;
       try
       {
-        final int nBCCount = aBusinessCardMgr.getSMPBusinessCardCount ();
+        final long nBCCount = aBusinessCardMgr.getSMPBusinessCardCount ();
 
         // Create new one
         aBusinessCard = aBusinessCardMgr.createOrUpdateSMPBusinessCard (aSG, new CommonsArrayList <> (aEntity1));
